requirements.txt
mitmproxy==5.0.0
pycryptodomex
